refactor: Scanpy-like tl/pl package layout
- Add scevonet.tl (models, programs, validation, adata) and scevonet.pl (plots)
- Keep net/viz/programs/validation/adata as thin re-exports for compatibility
- Fix Sample binary targets with index-aligned cell_types_df; fix draw_network gene=str
