+++ scevonet/pl/__init__.py
@@ -1,0 +1,5 @@
+"""Plotting (Scanpy ``scanpy.pl`` style)."""
+
+from scevonet.pl._basic import draw_net, draw_network, finish_matplotlib_figure
+
+__all__ = ["draw_net", "draw_network", "finish_matplotlib_figure"]
